animation-timing/callback-handle.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AnimationTiming Test: FrameRequestCallback - valid callback handle</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/animation-timing/#Window-interface-extensions">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

  async_test(function(t) {
    var requestId = window.requestAnimationFrame(callback);

    function callback(time) {
      t.step(function() {
        assert_greater_than(requestId, 0, "callback handle is a integer greater than zero");
      });
      t.done();
    };
  }, "Check window.requestAnimationFrame can return a valid callback handle");

</script>
